removed ProfileController and renamed UserEditType to UserType Symfony Form
